+++ fpa/src/main/scala/fpa/Clowns.scala
@@ -1,0 +1,253 @@
+package fpa
+
+object Clowns:
+
+  enum Expr0:
+    case Val(i: Int)
+    case Add(e1: Expr0, e2: Expr0)
+
+  import Expr0.*
+
+
+  /** non-tail recursive evaluator */
+
+  def eval1(e: Expr0): Int =
+    e match
+      case Val(i)      => i
+      case Add(e1, e2) => eval1(e1) + eval1(e2)
+
+
+  /** tail recursive evaluator employing a stack */
+
+  type Stack = List[Either[Expr0, Int]]
+
+  def eval2(e: Expr0): Int =
+
+    def load(e: Expr0)(stk: Stack): Int =
+      e match
+        case Val(i)      => unload(i)(stk)
+        case Add(e1, e2) => load(e1)(Left(e2) :: stk)
+
+    def unload(v1: Int)(stk: Stack): Int =
+      stk match
+        case Nil               => v1
+        case Left(e2) :: rest  => load(e2)(Right(v1) :: rest)
+        case Right(v2) :: rest => unload(v1 + v2)(rest)
+
+    load(e)(List.empty)
+
+
+  /** generic data component kit */
+
+  case class K1[A,B](a: A)                               // constant
+  case class I1[A](a: A)                                 // identity
+  enum S1[L[_], R[_], A]:                                // sum type aka either
+    case L1[L[_], R[_], A](la: L[A]) extends S1[L, R, A]
+    case R1[L[_], R[_], B](ra: R[B]) extends S1[L, R, B]
+  import S1.*
+  case class P1[L[_], R[_], A](la: L[A], ra: R[A])       // product type aka tuple
+
+
+
+  /** eg. unit generically defined as a constant */
+
+  type One[A] = K1[Unit,A]
+
+
+  /** or eg. the option type generically defined as a sum type */
+
+  type Maybe[A] = S1[One,I1,A]
+
+  def nothing[A]: Maybe[A] =
+    L1(K1(()))
+
+  def just[A](a: A): Maybe[A] =
+    R1(I1(a))
+
+
+  /** the kit is functorial */
+
+  trait Functor[F[_]]:
+    def fmap[A,B](f: A => B)(fa: F[A]): F[B]
+
+  implicit def kFunctor[X]: Functor[K1[X, *]] =
+    new Functor[K1[X, *]]:
+      def fmap[A, B](f: A => B)(fa: K1[X, A]): K1[X, B] = K1(fa.a)
+
+  implicit def iFunctor: Functor[I1[_]] =
+    new Functor[I1[_]]:
+      def fmap[A, B](f: A => B)(fa: I1[A]): I1[B] = I1(f(fa.a))
+
+  implicit def sFunctor[L[_], R[_]](implicit lFunctor: Functor[L], rFunctor: Functor[R]): Functor[S1[L, R, _]] =
+    new Functor[S1[L, R, _]]:
+      def fmap[A, B](f: A => B)(fa: S1[L, R, A]): S1[L, R, B] =
+        fa match
+          case L1(la) => L1(lFunctor.fmap(f)(la))
+          case R1(ra) => R1(rFunctor.fmap(f)(ra))
+
+  implicit def pFunctor[L[_], R[_]](implicit lFunctor: Functor[L], rFunctor: Functor[R]): Functor[P1[L, R, _]] =
+    new Functor[P1[L, R, _]]:
+      def fmap[A, B](f: A => B)(fa: P1[L, R, A]): P1[L, R, B] =
+        P1(lFunctor.fmap(f)(fa.la), rFunctor.fmap(f)(fa.ra))
+
+
+  /** the expr branching structure is readily described by a polynomial */
+  type ExprP[A] = S1[K1[Int, *], P1[I1, I1, *], A]
+
+  def valP[A](i: Int): ExprP[A] =
+    L1(K1(i))
+
+  object ValP:
+    def unapply(ep: ExprP[Int]): Option[Int] =
+      ep match
+        case L1(K1(i)) => Some(i)
+        case _         => None
+
+  def addP[A](e1: A, e2: A): ExprP[A] =
+    R1(P1(I1(e1), I1(e2)))
+
+  object AddP:
+    def unapply(ep: ExprP[Int]): Option[(Int, Int)] =
+      ep match
+        case R1(P1(I1(e1), I1(e2))) => Some(e1, e2)
+        case _                      => None
+
+
+  /** we would like now to establish the isomorphism: Expr ∼= ExprP Expr */
+  /** which we do via a type level fix point to tie the knot inductively */
+  case class Mu[F[_]](in: F[Mu[F]])
+
+  type Expr = Mu[ExprP]
+
+  def valM(i: Int): Expr =
+    Mu(valP(i))
+
+  def addM(e1: Expr, e2: Expr): Expr =
+    Mu(addP(e1, e2))
+
+
+  /** this lets us define a fold like recursion operator as a catamorphism - see doc/img/cata.png */
+  def cata[P[_], A](phi: P[A] => A)(p: Mu[P])(implicit pFunctor: Functor[P]): A =
+    phi(pFunctor.fmap(cata(phi))(p.in))
+
+  /** with a subsequent evaluator in terms of that catamorphism */
+  def  eval3(e: Expr): Int =
+    def phi(e: ExprP[Int]): Int =
+      e match
+        case ValP(i)      => i
+        case AddP(e1, e2) => e1 + e2
+        case _            => sys.error("boom")
+    cata(phi)(e)
+
+
+  /**
+    * we shall see how to turn a cata into a first-order tail-recursion whenever p is polynomial
+    * we shall do this by dissecting p, with ‘clown’ elements left and ‘joker’s on the right
+    * to this end, we need polynomial bifunctors, which are just functors, but in two directions
+    * let's construct a data component kit for that with two generic type parameters `x` and `y`
+    * generic data component kit
+    **/
+  case class K2[X, A, B](const: X)                                       // constant
+  case class Fst[A, B](a: A)                                             // first identity
+  case class Snd[A, B](b: B)                                             // second identity
+  enum S2[L[_, _], R[_, _], A, B]:                                       // sum type aka either
+    case L2[L[_, _], R[_, _], A, B](la: L[A, B]) extends S2[L, R, A, B]
+    case R2[L[_, _], R[_, _], A, B](ra: R[A, B]) extends S2[L, R, A, B]
+  import S2.*
+  case class P2[L[_, _], R[_, _], A, B](la: L[A, B], ra: R[A, B])        // product type aka tuple
+
+
+  /** and yes - the kit is bifunctorial */
+  trait Bifunctor[F[_,_]]:
+    def bimap[A, B, C, D](f: A => C)(g: B => D)(fab: F[A, B]): F[C, D]
+
+  implicit def k2Functor[X]: Bifunctor[K2[X, *, *]] =
+    new Bifunctor[K2[X, *, *]]:
+      def bimap[A, B, C, D](f: A => C)(g: B => D)(fab: K2[X, A, B]): K2[X, C, D] = K2(fab.const)
+
+  implicit def fstFunctor: Bifunctor[Fst[*, *]] =
+    new Bifunctor[Fst[*, *]]:
+      def bimap[A, B, C, D](f: A => C)(g: B => D)(fab: Fst[A, B]): Fst[C, D] = Fst(f(fab.a))
+
+  implicit def sndFunctor: Bifunctor[Snd[*, *]] =
+    new Bifunctor[Snd[*, *]]:
+      def bimap[A, B, C, D](f: A => C)(g: B => D)(fab: Snd[A, B]): Snd[C, D] = Snd(g(fab.b))
+
+  implicit def s2Functor[L[_, _], R[_, _]](implicit lFunctor: Bifunctor[L], rFunctor: Bifunctor[R]): Bifunctor[S2[L, R, *, *]] =
+    new Bifunctor[S2[L, R, *, *]]:
+      def bimap[A, B, C, D](f: A => C)(g: B => D)(fab: S2[L, R, A, B]): S2[L, R, C, D] =
+        fab match
+          case L2(la) => L2(lFunctor.bimap(f)(g)(la))
+          case R2(rb) => R2(rFunctor.bimap(f)(g)(rb))
+
+  implicit def p2Functor[L[_, _], R[_, _]](implicit lFunctor: Bifunctor[L], rFunctor: Bifunctor[R]): Bifunctor[P2[L, R, *, *]] =
+    new Bifunctor[P2[L, R, *, *]]:
+      def bimap[A, B, C, D](f: A => C)(g: B => D)(fab: P2[L, R, A, B]): P2[L, R, C, D] =
+        P2(lFunctor.bimap(f)(g)(fab.la), rFunctor.bimap(f)(g)(fab.ra))
+
+
+  /** but.. but.. nothing is missing - we need non-constructable zero */
+  sealed trait Zero
+
+  def magic[A](z: Zero): A =
+    sys.error("we never get this far")
+
+  /** still, nothing is functorial considering F[Zero] as a container with zero elements */
+  def inflate1[F[_], A](fz: F[Zero])(implicit zFunctor: Functor[F]): F[A] =
+    zFunctor.fmap(magic)(fz)
+
+  def inflate2[F[_], A](fz: F[Zero])(implicit  zFunctor: Functor[F]): F[A] =
+    fz.asInstanceOf[F[A]]
+
+  /** and nothing can be defined in terms of our functorial and bifuncorial kit as the constant zero */
+  type Zero1[A]    = K1[Zero, A]
+  type Zero2[A, B] = K2[Zero, A, B]
+
+  /** with that we can dissect left clowns and right jokers */
+  case class Clown[F[_], C, J](fc: F[C])
+
+  implicit def clownBifunctor[F[_]](implicit functor: Functor[F]): Bifunctor[Clown[F, *, *]] =
+    new Bifunctor[Clown[F, *, *]]:
+      override def bimap[A, B, C, D](f: A => C)(g: B => D)(fab: Clown[F, A, B]): Clown[F, C, D] =
+        Clown(functor.fmap(f)(fab.fc))
+
+  /** note that I1 ∼= Fst. */
+
+  case class Joker[F[_], C, J](fj: F[J])
+
+  implicit def jokerBifunctor[F[_]](implicit functor: Functor[F]): Bifunctor[Joker[F, *, *]] =
+    new Bifunctor[Joker[F, *, *]]:
+      override def bimap[A, B, C, D](f: A => C)(g: B => D)(fab: Joker[F, A, B]): Joker[F, C, D] =
+        Joker(functor.fmap(g)(fab.fj))
+
+  /** given a left functor and right bifunctor we define a dissection to have an implication on the right  */
+  /** with that we can define instances in terms of our left and right component kits of which the product */
+  /** type proves to require the most horrific encoding                                                    */
+
+
+  object Main:
+
+    @main
+    def runClowns(args: String*): Unit =
+
+      val e0: Expr0 = Add(Add(Val(1), Val(2)), Val(3))
+
+      val a1 = eval1(e0)
+      val a2 = eval2(e0)
+      println(s"eval1($e0)=$a1")
+      println(s"eval2($e0)=$a2")
+      assert(a1 == 6)
+      assert(a2 == 6)
+
+      val m1 = implicitly[Functor[Maybe]].fmap[Char,Boolean](_.isDigit)(just('1'))
+      val m2 = implicitly[Functor[Maybe]].fmap[Char,Boolean](_.isDigit)(nothing[Char])
+      println(s"implicitly[Functor[Maybe]].fmap[Char,Boolean](_.isDigit)(just('1'))=$m1")
+      println(s"implicitly[Functor[Maybe]].fmap[Char,Boolean](_.isDigit)(nothing[Char])=$m2")
+      assert(m1 == R1(I1(true)))
+      assert(m2 == L1(K1(())))
+
+      val e: Expr = addM(addM(valM(1), valM(2)), valM(3))
+
+      val a3 = eval3(e)
+      println(s"eval3($e)=$a3")
+      assert(a2 == 6)
