+++ fpa/src/main/scala/fpa/Properties.scala
@@ -1,6 +1,6 @@
 package fpa
 
 /** All exercises are defined in `test/scala/fpa/properties.scala` */
-object placeholder
+object Properties
