--- fpa/src/main/scala/fpa/intro.scala
@@ -1,12 +1,0 @@
-package fpa
-package intro
-
-// A version of the archetypical, introductionary "Hello World" example in Scala
-
-object Main extends Greeting with App {
-  println(greeting)
-}
-
-trait Greeting {
-  lazy val greeting: String = "Hello Anonymous Functional Programmer !!!"
-}
